Fix part of PR78555 - gcc/real.c:2890:25: runtime error: left shift of negative value -125
	PR ipa/78555
	* real.c (real_hash): Add cast to avoid left
	shifting of negative values.

From-SVN: r243012

+++ gcc/real.c
@@ -2887,7 +2887,7 @@ real_hash (const REAL_VALUE_TYPE *r)
       return h;
 
     case rvc_normal:
-      h |= REAL_EXP (r) << 3;
+      h |= (unsigned int)REAL_EXP (r) << 3;
       break;
 
     case rvc_nan:
